Refactor: 매직 리터럴,매직넘버 처리
- 매직 리터럴과 매직넘버를 상수로 변환시킴.
- 예외처리 메시지같은 경우는 각 Exception마다 공통된 메시지를 넣어주는 로직인 occur 메서드를 추가함.
- 여러곳에서 사용되는 상수는 별도의 클래스를 생성해줌.
