+++ src/main/java/christmas/domain/UserInformation.java
@@ -1,6 +1,5 @@
 package christmas.domain;
 
-import christmas.util.GiveawayEvent;
 import java.util.Map;
 
 public class UserInformation {
