+++ src/main/java/christmas/domain/constant/Constant.java
@@ -1,0 +1,5 @@
+package christmas.domain.constant;
+
+public class Constant {
+    public static final String GIVEAWAY_EVENT_TITLE = "증정 이벤트";
+}
