+++ src/main/java/christmas/exception/VisitedDateException.java
@@ -1,6 +1,7 @@
 package christmas.exception;
 
 public class VisitedDateException extends IllegalArgumentException {
+    private static final String EXCEPTION_MESSAGE = "[ERROR] 유효하지 않은 날짜입니다. 다시 입력해 주세요.";
 
     private VisitedDateException(String exceptionMessage) {
         super(exceptionMessage);
@@ -8,5 +9,9 @@ public class VisitedDateException extends IllegalArgumentException {
 
     public static VisitedDateException from(String exceptionMessage) {
         return new VisitedDateException(exceptionMessage);
+    }
+
+    public static VisitedDateException occur() {
+        return new VisitedDateException(EXCEPTION_MESSAGE);
     }
 }
