+++ src/main/java/christmas/util/MenuValidator.java
@@ -15,19 +15,19 @@ public class MenuValidator {
 
     private static void menuExistValidate(String menu) {
         if (Menu.findMenu(menu) == Menu.NONE) {
-            throw MenuException.from("[ERROR] 유효하지 않은 주문입니다. 다시 입력해 주세요.");
+            throw MenuException.occur();
         }
     }
 
     private static void amountNumberValidate(String amount) {
         if (Validator.isNotNumber(amount)) {
-            throw MenuException.from("[ERROR] 유효하지 않은 주문입니다. 다시 입력해 주세요.");
+            throw MenuException.occur();
         }
     }
 
     private static void amountMinimunValidate(int amount) {
         if (amount < MINIMUM_ORDER_AMOUNT) {
-            throw MenuException.from("[ERROR] 유효하지 않은 주문입니다. 다시 입력해 주세요.");
+            throw MenuException.occur();
         }
     }
